Started making login processing all on one file
Turns out the only problem i had the first time i tried to do it was I
used <? instead of <?php. I've applied the change to the index page, and
it looks like it works so i'm going to start changing all pages.
I've also been cleaning up the admin pages, and will eventually combine
all of them as well.