Cleanup JS
moved the header functions in js to a separate file. So far the only
functions are headerFunctions, which has everything needed for the
create account stuff, and hideCreateAccount which hides the ability to
create an account. If a variable is passed into the function, it also
hides the log in as org ability. Also fixed css for the direct create
account link and removed the type attribute from most js tags