Started extending function of announcement interactivity
Added options in creation menu to make the type of the announcement be 1
of 3, either only text, a url, or modal popup. Right now all it does is
show the option to input a url if the link option is selected. It also
makes this option required. Type is inputted into the database, but it
doesn't do anything yet. URL is not entered into database yet.